hamyaran/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
cryptography==41.0.7
Django==5.1.4
django-crispy-forms==2.3
django-crontab==0.7.1
django-dbbackup==4.2.1
django-gravatar==0.1.0
django-gravatar2==1.4.5
django-jalali==7.2.0
django-push-notifications==3.1.0
django-pwa==2.0.1
django-webpush==0.3.6
django-widget-tweaks==1.5.0
dnspython==2.7.0
email_validator==2.2.0
frozenlist==1.5.0
http_ece==1.2.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.4.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
py-vapid==1.9.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywebpush==2.0.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webpush==1.0.3
yarl==1.18.3
